requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
rawpy==0.26.1
scipy==1.13.0
aiofiles==24.1.0
numpy>=2.0.0
opencv-python-headless>=4.10.0
Pillow>=10.0.0
pydantic>=2.0.0
